# Convert Attribute to a struct