@ramonza 2018-03-14
Not sure why the `delta` column displays just "~" but you can see that the numbers did change. This test compares the numbers after making everything sampled (before, we didn't benchmark the sampled case, I also added sampled/not-sampled benchmarks).