@odeke-em 2018-03-14
>Not sure why the delta column displays just "~" but you can see that the numbers did change

If you didn't already, could you please run the benchmarks say 10 times or so e.g
```shell
go test -run=^$ -bench=. -count=10
```