@ramonza 2018-03-14
Unfortunately it isn't easy to reproduce the above table with "count=10" because I changed the benchmarks themselves in this PR. In any case, the performance gain is not the only reason for this change - it is more idiomatic (returning structs rather than interfaces) and it's a net code reduction (except for more test code).